Añadidos campos a entity y Dto

+++ src/main/java/com/init/Jobarcelona/repository/ChampRepository.java
@@ -3,5 +3,5 @@ package com.init.Jobarcelona.repository;
 import com.init.Jobarcelona.entity.Champ;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ChampRepository extends JpaRepository<Champ, Integer> {
+public interface ChampRepository extends JpaRepository<Champ, Long> {
 }
